Examples: Move categorization to examplesparser
For easier testing.
