+++ src/plugins/qtsupport/examplesparser.cpp
@@ -3,7 +3,10 @@
 
 #include "examplesparser.h"
 
+#include "qtsupporttr.h"
+
 #include <utils/algorithm.h>
+#include <utils/environment.h>
 #include <utils/filepath.h>
 #include <utils/stylehelper.h>
 
@@ -296,6 +299,57 @@ expected_str<QList<ExampleItem *>> parseExamples(const QByteArray &manifestData,
                                    .arg(reader.errorString()));
     }
     return items;
+}
+
+static bool sortByHighlightedAndName(ExampleItem *first, ExampleItem *second)
+{
+    if (first->isHighlighted && !second->isHighlighted)
+        return true;
+    if (!first->isHighlighted && second->isHighlighted)
+        return false;
+    return first->name.compare(second->name, Qt::CaseInsensitive) < 0;
+}
+
+QList<std::pair<QString, QList<ExampleItem *>>> getCategories(const QList<ExampleItem *> &items,
+                                                              bool sortIntoCategories)
+{
+    static const QString otherDisplayName = Tr::tr("Other", "Category for all other examples");
+    const bool useCategories = sortIntoCategories
+                               || qtcEnvironmentVariableIsSet("QTC_USE_EXAMPLE_CATEGORIES");
+    QList<ExampleItem *> other;
+    QMap<QString, QList<ExampleItem *>> categoryMap;
+    if (useCategories) {
+        for (ExampleItem *item : items) {
+            const QStringList itemCategories = item->metaData.value("category");
+            for (const QString &category : itemCategories)
+                categoryMap[category].append(item);
+            if (itemCategories.isEmpty())
+                other.append(item);
+        }
+    }
+    QList<std::pair<QString, QList<ExampleItem *>>> categories;
+    if (categoryMap.isEmpty()) {
+        // The example set doesn't define categories. Consider the "highlighted" ones as "featured"
+        QList<ExampleItem *> featured;
+        QList<ExampleItem *> allOther;
+        std::tie(featured, allOther) = Utils::partition(items, [](ExampleItem *i) {
+            return i->isHighlighted;
+        });
+        if (!featured.isEmpty())
+            categories.append({Tr::tr("Featured", "Category for highlighted examples"), featured});
+        if (!allOther.isEmpty())
+            categories.append({otherDisplayName, allOther});
+    } else {
+        const auto end = categoryMap.constKeyValueEnd();
+        for (auto it = categoryMap.constKeyValueBegin(); it != end; ++it)
+            categories.append(*it);
+        if (!other.isEmpty())
+            categories.append({otherDisplayName, other});
+    }
+    const auto end = categories.end();
+    for (auto it = categories.begin(); it != end; ++it)
+        sort(it->second, sortByHighlightedAndName);
+    return categories;
 }
 
 } // namespace QtSupport::Internal
